perf(api): gzip responses, fast map-pins serialization, cache headers (#142)
- GZipMiddleware (min 1 KiB, level 6): compresses the SPA bundle and
  large JSON like /api/map-pins (~1 MB -> ~150 KB)
- /api/map-pins: hand-rolled dicts + direct JSONResponse instead of
  double Pydantic validation of 10k rows; lat/lon now real JSON numbers
  rounded to 5 dp (~1 m) instead of Decimal-as-string, matching the
  frontend's declared lat: number type
- Cache-Control: map-pins public/max-age=300; hashed /assets/ immutable
  for a year; rendered index.html no-cache so deploys are picked up
- drop duplicate seo.router include

+++ warn_v2/api/schemas.py
@@ -182,5 +182,5 @@ class MapPinOut(BaseModel):
     state: str
     notice_date: date | None
     layoff_count: int | None
-    lat: Decimal
-    lon: Decimal
+    lat: float
+    lon: float
